Einops

Documentation for Einops.

Einops.pack

Einops.rearrange

Einops.unpack

Einops.@einops_str

Einops.pack — Method

pack(unpacked_arrays, pattern)

Pack a vector of arrays into a single array according to the pattern.

Examples

```jldoctest julia> inputs = [rand(2,3,5), rand(2,3,7,5), rand(2,3,7,9,5)]

julia> packedarray, packedshapes = pack(inputs, (:i, :j, *, :k));

julia> size(packed_array) (2, 3, 71, 5)

julia> packed_shapes 3-element Vector{NTuple{N, Int64} where N}: () (7,) (7, 9)

Einops.rearrange — Method

rearrange(x::AbstractArray, left --> right)

Rearrange the axes of x according to the pattern specified by left --> right.

Can always be expressed as a reshape + permutedims + reshape.

Examples

julia> rearrange(rand(2,3,5), (:a, :b, :c) --> (:c, :b, :a)) |> size
(5, 3, 2)

julia> permutedims(rand(2,3,5), (3,2,1)) |> size
(5, 3, 2)

julia> rearrange(rand(2,3,35), (:a, :b, (:c, :d)) --> (:a, :d, (:c, :b)), c=5) |> size
(2, 7, 15)

julia> reshape(permutedims(reshape(rand(2,3,35), 2,3,5,7), (1,4,3,2)), 2,7,5*3) |> size
(2, 7, 15)

Einops.unpack — Method

unpack(packed_array, packed_shapes, pattern)

Unpack a single array into a vector of arrays according to the pattern.

Examples

julia> inputs = [rand(2,3,5), rand(2,3,7,5), rand(2,3,7,9,5)];

julia> inputs == unpack(pack(inputs, (:i, :j, *, :k))..., (:i, :j, *, :k))
true

julia> packed_array = rand(2,3,16);

julia> packed_shapes = [(), (7,), (4, 2)];

julia> unpack(packed_array, packed_shapes, (:i, :j, *)) .|> size
3-element Vector{Tuple{Int64, Int64, Vararg{Int64}}}:
 (2, 3)
 (2, 3, 7)
 (2, 3, 4, 2)

Einops.@einops_str — Macro

einops"a (b c) -> (c b a)"
einops"i j * k"

For parity with Python implementation.

Examples

julia> einops"a 1 b c -> (c b) a"
(:a, 1, :b, :c) --> ((:c, :b), :a)

julia> einops"embed token (head batch) -> (embed head) token batch"
(:embed, :token, (:head, :batch)) --> ((:embed, :head), :token, :batch)

julia> einops"i j * k" # for packing
(:i, :j, *, :k)